Added Bank Operations and Stock buy

+++ Services/layer1/Order_Service.py
@@ -1,5 +1,5 @@
 from Services.layer2 import Stock_Service, Fund_Service
-
+from flask import redirect, render_template
 
 def buy(ticker, quantity, person):
     """ Service to buy a stock
@@ -11,11 +11,14 @@ def buy(ticker, quantity, person):
         Success or Error Page """
     stock_price = Stock_Service.get_stock_price(ticker)
     total_price = quantity * stock_price
+
+    """ if insufficient funds render warning """
+    if person.balance < total_price:
+        return render_template('buy_stock.html', ticker=ticker, stock_price=stock_price, failure=True, cost=total_price, cash_value=person.balance)
+
     Fund_Service.remove_funds(total_price, person)
-    # TODO
-    """ if user has enough money, subtract total_price from user balance and add buy transaction to transactions table"""
-    """ else: return error"""
-    return '/stocks/show/' + ticker + '/'
+    # TODO add buy transaction to transactions table and holdings table"""
+    return redirect('/stocks/show/' + ticker + '/')
 
 
 def sell(ticker, quantity, person):
@@ -29,7 +32,9 @@ def sell(ticker, quantity, person):
     stock_price = Stock_Service.get_stock_price(ticker)
     total_price = quantity * stock_price
     Fund_Service.add_funds(total_price, person)
-    # TODO
+    # TODO check they own the stocks and quantity, remove from holdings, add to transactions
     """ if user has enough stocks to sell, add total_price to user balance and add sell transaction to Transactions table"""
     """ else: return error"""
     return '/stocks/show/' + ticker + '/'
+
+
